+++ src/test/integration/java/io/github/jaredpetersen/kafkaconnectredis/sink/RedisSinkTaskIT.java
@@ -15,6 +15,7 @@ import java.util.Map;
 import org.apache.kafka.connect.data.Schema;
 import org.apache.kafka.connect.data.SchemaBuilder;
 import org.apache.kafka.connect.data.Struct;
+import org.apache.kafka.connect.errors.ConnectException;
 import org.apache.kafka.connect.sink.SinkRecord;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
@@ -29,6 +30,7 @@ import org.testcontainers.utility.MountableFile;
 import reactor.test.StepVerifier;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @Testcontainers
 public class RedisSinkTaskIT {
@@ -206,6 +208,17 @@ public class RedisSinkTaskIT {
         .create(REDIS_CLUSTER_COMMANDS.dbsize())
         .expectNext(0L)
         .verifyComplete();
+  }
+
+  @Test
+  public void startThrowsConnectExceptionForInvalidConfig() {
+    final RedisSinkTask sinkTask = new RedisSinkTask();
+
+    final Map<String, String> connectorConfig = new HashMap<>();
+    connectorConfig.put("redis.cluster.enabled", "false");
+
+    final ConnectException thrown = assertThrows(ConnectException.class, () -> sinkTask.start(connectorConfig));
+    assertEquals("task configuration error", thrown.getMessage());
   }
 
   @Test