+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2020-12-11
+### Added
+- Parallelization for source connector based on channels/patterns
+
+### Removed
+- Default configuration for Kafka topic
+
 ## [1.0.4] - 2020-11-29
 ### Added
 - Added support for sinking arbitrary Redis commands, primarily for use with Redis modules